+++ packages/cli/examples/delegate.ts
@@ -28,13 +28,8 @@ const { accountNumber, sequence } = await client.getSequence(senderAddress);
 console.log("Account/sequence:", accountNumber, sequence);
 
 const signDoc = makeSignDoc([msg], fee, chainId, memo, accountNumber, sequence);
-const { signature } = await wallet.sign(senderAddress, signDoc);
-const signedTx: StdTx = {
-  msg: [msg],
-  fee: fee,
-  memo: memo,
-  signatures: [signature],
-};
+const { signed, signature } = await wallet.sign(senderAddress, signDoc);
+const signedTx = makeStdTx(signed, signature);
 
 const result = await client.broadcastTx(signedTx);
 console.log("Broadcast result:", result);