+++ packages/cosmwasm/src/cosmwasmclient.spec.ts
@@ -5,10 +5,10 @@ import {
   assertIsBroadcastTxSuccess,
   isWrappedStdTx,
   makeSignDoc,
+  makeStdTx,
   MsgSend,
   Secp256k1Wallet,
   StdFee,
-  StdTx,
 } from "@cosmjs/launchpad";
 import { assert, sleep } from "@cosmjs/utils";
 import { ReadonlyDate } from "readonly-date";
@@ -240,13 +240,8 @@ describe("CosmWasmClient", () => {
       const chainId = await client.getChainId();
       const { accountNumber, sequence } = await client.getSequence(alice.address0);
       const signDoc = makeSignDoc([sendMsg], fee, chainId, memo, accountNumber, sequence);
-      const { signature } = await wallet.sign(alice.address0, signDoc);
-      const signedTx: StdTx = {
-        msg: [sendMsg],
-        fee: fee,
-        memo: memo,
-        signatures: [signature],
-      };
+      const { signed, signature } = await wallet.sign(alice.address0, signDoc);
+      const signedTx = makeStdTx(signed, signature);
       const result = await client.broadcastTx(signedTx);
       assertIsBroadcastTxSuccess(result);
       const { logs, transactionHash } = result;