+++ packages/cosmwasm/src/signingcosmwasmclient.ts
@@ -12,11 +12,11 @@ import {
   GasPrice,
   isBroadcastTxFailure,
   makeSignDoc,
+  makeStdTx,
   Msg,
   MsgSend,
   OfflineSigner,
   StdFee,
-  StdTx,
 } from "@cosmjs/launchpad";
 import { Uint53 } from "@cosmjs/math";
 import pako from "pako";
@@ -361,13 +361,8 @@ export class SigningCosmWasmClient extends CosmWasmClient {
     const { accountNumber, sequence } = await this.getSequence();
     const chainId = await this.getChainId();
     const signDoc = makeSignDoc(msgs, fee, chainId, memo, accountNumber, sequence);
-    const { signature } = await this.signer.sign(this.senderAddress, signDoc);
-    const signedTx: StdTx = {
-      msg: msgs,
-      fee: fee,
-      memo: memo,
-      signatures: [signature],
-    };
+    const { signed, signature } = await this.signer.sign(this.senderAddress, signDoc);
+    const signedTx = makeStdTx(signed, signature);
     return this.broadcastTx(signedTx);
   }
 }