+++ packages/cosmwasm/src/testutils.spec.ts
@@ -1,6 +1,5 @@
 import { Random } from "@cosmjs/crypto";
 import { Bech32, fromBase64 } from "@cosmjs/encoding";
-import { Msg, StdFee, StdSignature, StdTx } from "@cosmjs/launchpad";
 
 import hackatom from "./testdata/contract.json";
 
@@ -88,13 +87,4 @@ export function pendingWithoutWasmd(): void {
 export function fromOneElementArray<T>(elements: ArrayLike<T>): T {
   if (elements.length !== 1) throw new Error(`Expected exactly one element but got ${elements.length}`);
   return elements[0];
-}
-
-export function makeSignedTx(firstMsg: Msg, fee: StdFee, memo: string, firstSignature: StdSignature): StdTx {
-  return {
-    msg: [firstMsg],
-    fee: fee,
-    memo: memo,
-    signatures: [firstSignature],
-  };
 }