+++ packages/launchpad/src/testutils.spec.ts
@@ -1,10 +1,6 @@
 import { Random } from "@cosmjs/crypto";
 import { Bech32 } from "@cosmjs/encoding";
 
-import { Msg } from "./msgs";
-import { StdTx } from "./tx";
-import { StdFee, StdSignature } from "./types";
-
 export function makeRandomAddress(): string {
   return Bech32.encode("cosmos", Random.getBytes(20));
 }
@@ -74,13 +70,4 @@ export function pendingWithoutWasmd(): void {
 export function fromOneElementArray<T>(elements: ArrayLike<T>): T {
   if (elements.length !== 1) throw new Error(`Expected exactly one element but got ${elements.length}`);
   return elements[0];
-}
-
-export function makeSignedTx(firstMsg: Msg, fee: StdFee, memo: string, firstSignature: StdSignature): StdTx {
-  return {
-    msg: [firstMsg],
-    fee: fee,
-    memo: memo,
-    signatures: [firstSignature],
-  };
 }